+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fix `scripts-lint` on `watch` that would stop on lint error
+- Fix `styles-lint` on `watch` that would still halt after fixing lint error, by removing custom patch